clean up old deprecated files mark project plan as obsolete
