--- src/Depreciated/CrawlerCollection.cs
@@ -1,20 +1,0 @@
-// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
-// The Apache v2. See License.txt in the project root for license information.
-
-using System;
-
-namespace Wangkanai.Detection
-{
-    [Obsolete]
-    internal static class CrawlerCollection
-    {
-        public static string[] Keywords => new string[] {
-            "bot",
-            "slurp",
-            "spider",
-            "google-structured-data-testing-tool",
-            "facebookexternalhit",
-            "skypeuripreview"
-        };
-    }
-}
