+++ src/Depreciated/IDeviceFactory.cs
@@ -2,12 +2,14 @@
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
+using Wangkanai.Detection.Models;
 
 namespace Wangkanai.Detection
 {
     [Obsolete]
-    public interface IPlatformResolver : IResolver
+    public interface IDeviceFactory
     {
-        IPlatformFactory Platform { get; }
+        Device Type { get; set; }
+        bool Crawler { get; set; }
     }
 }
